refactor(cli): extract command dispatch and command-file module (#160)
